pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "genshin_mummy"
version = "0.0.1"
authors = [{ name = "LetianLiu", email = "15261172997@163.com" }]
description = "A tech-neet's RPA project for Genshin Impact."
readme = "README.md"
license = { file = "LICENSE" }
classifiers = ["License :: OSI Approved :: MIT License"]
dependencies = [
    "attrs~=23.1.0",
    "iolite~=22.2.0",
    "numpy~=1.24.4",
    "openpyxl~=3.1.2",
    "pyautogui~=0.9.54",
    "pillow~=10.1.0",
    "pyscreeze~=0.1.30",
    "opencv-python~=4.6.0.66",
    "paddlepaddle~=2.5.2",
    "paddleocr~=2.7.0.3",
]

[project.optional-dependencies]
dev = [
    "yapf"
]

[project.scripts]
fuck-shit-artifact = "genshin_mummy.artifact_helper.unlock_shit_artifact:main"
